Install BP4O init scripts in /etc/profile.d/
The /etc/profile.d/ directory is used in Bash as a place to put scripts that need to be sourced
at login.  Since BP4O is a script that needs to be sourced at login, /etc/profile.d/ is a perfect
place for BP4O's init scripts!  This means that root installations for Bash users won't need to
source anything in their .bashrc file.

For non-root installations, BP4O init scripts will be placed in ~/.profile.d/ and the user will
need to source the BP4O init scripts.

Zsh users, depending on their distro, will need to source the BP4O init scripts regardless of
whether they install as root or not.