Fish: Fix improper string handling
bp4o would fail if you ran 'p4 -V' to find the version of your p4, because the string
function read the '-V' as an option instead of an input.
Added the -- option to all string function calls, which specifies that there are no more options
past this.

bp4o was also incorrectly checking the returned alias var, and would fail when running a
non-alias command.