Allow tests to run from extracted tarball
In each of test.sh and test/*.*sh, the cwd variable is set by running
`git rev-parse --show-toplevel`. While this works fine from a git clone,
it breaks if the tests are run from an extracted tarball - either we
aren't in a git repo or we get the wrong toplevel directory.

To fix this, in test.sh, we set the cwd variable to the absolute path of
the directory containing the test.sh script and cd into that directory
before sourcing any of the test/*.*sh files. In the test/*.*sh files,
the cwd variable is simply determined by running the cwd command, which
returns the directory we cded into before.

Tests run:
    - Ran `./test.sh` from within a git clone. All the tests passed.
    - Downloaded and extracted the v1.1.1.tar.gz, applied this commit as
      a patch to the result, and ran `./test.sh`. All the tests now pass
      in this context.